Use sample instead of dummy
