+++ test/plugin/test_in_sample.rb
@@ -1,33 +1,33 @@
 require_relative '../helper'
 require 'fluent/test/driver/input'
-require 'fluent/plugin/in_dummy'
+require 'fluent/plugin/in_sample'
 require 'fileutils'
 
-class DummyTest < Test::Unit::TestCase
+class SampleTest < Test::Unit::TestCase
   def setup
     Fluent::Test.setup
   end
 
   def create_driver(conf)
-    Fluent::Test::Driver::Input.new(Fluent::Plugin::DummyInput).configure(conf)
+    Fluent::Test::Driver::Input.new(Fluent::Plugin::SampleInput).configure(conf)
   end
 
   sub_test_case 'configure' do
     test 'required parameters' do
       assert_raise_message("'tag' parameter is required") do
-        Fluent::Plugin::DummyInput.new.configure(config_element('ROOT',''))
+        Fluent::Plugin::SampleInput.new.configure(config_element('ROOT',''))
       end
     end
 
     test 'tag' do
       d = create_driver(%[
-        tag dummy
+        tag sample
       ])
-      assert_equal "dummy", d.instance.tag
+      assert_equal "sample", d.instance.tag
     end
 
     config = %[
-      tag dummy
+      tag sample
     ]
 
     test 'auto_increment_key' do
@@ -44,30 +44,30 @@ class DummyTest < Test::Unit::TestCase
       assert_equal 10, d.instance.rate
     end
 
-    test 'dummy' do
+    test 'sample' do
       # hash is okay
-      d = create_driver(config + %[dummy {"foo":"bar"}])
-      assert_equal [{"foo"=>"bar"}], d.instance.dummy
+      d = create_driver(config + %[sample {"foo":"bar"}])
+      assert_equal [{"foo"=>"bar"}], d.instance.sample
 
       # array of hash is okay
-      d = create_driver(config + %[dummy [{"foo":"bar"}]])
-      assert_equal [{"foo"=>"bar"}], d.instance.dummy
+      d = create_driver(config + %[sample [{"foo":"bar"}]])
+      assert_equal [{"foo"=>"bar"}], d.instance.sample
 
       assert_raise_message(/JSON::ParserError|got incomplete JSON/) do
-        create_driver(config + %[dummy "foo"])
+        create_driver(config + %[sample "foo"])
       end
 
       assert_raise_message(/is not a hash/) do
-        create_driver(config + %[dummy ["foo"]])
+        create_driver(config + %[sample ["foo"]])
       end
     end
   end
 
   sub_test_case "emit" do
     config = %[
-      tag dummy
+      tag sample
       rate 10
-      dummy {"foo":"bar"}
+      sample {"foo":"bar"}
     ]
 
     test 'simple' do
@@ -75,7 +75,7 @@ class DummyTest < Test::Unit::TestCase
       d.run(timeout: 0.5)
 
       d.events.each do |tag, time, record|
-        assert_equal("dummy", tag)
+        assert_equal("sample", tag)
         assert_equal({"foo"=>"bar"}, record)
         assert(time.is_a?(Fluent::EventTime))
       end
@@ -86,20 +86,20 @@ class DummyTest < Test::Unit::TestCase
       d.run(timeout: 0.5)
 
       d.events.each_with_index do |(tag, _time, record), i|
-        assert_equal("dummy", tag)
+        assert_equal("sample", tag)
         assert_equal({"foo"=>"bar", "id"=>i}, record)
       end
     end
   end
 
-  TEST_PLUGIN_STORAGE_PATH = File.join( File.dirname(File.dirname(__FILE__)), 'tmp', 'in_dummy', 'store' )
+  TEST_PLUGIN_STORAGE_PATH = File.join( File.dirname(File.dirname(__FILE__)), 'tmp', 'in_sample', 'store' )
   FileUtils.mkdir_p TEST_PLUGIN_STORAGE_PATH
 
-  sub_test_case 'when dummy plugin has storage which is not specified the path'  do
+  sub_test_case 'when sample plugin has storage which is not specified the path'  do
     config1 = {
-      'tag' => 'dummy',
+      'tag' => 'sample',
       'rate' => '0',
-      'dummy' => '[{"x": 1, "y": "1"}, {"x": 2, "y": "2"}, {"x": 3, "y": "3"}]',
+      'sample' => '[{"x": 1, "y": "1"}, {"x": 2, "y": "2"}, {"x": 3, "y": "3"}]',
       'auto_increment_key' => 'id',
     }
     conf1 = config_element('ROOT', '', config1, [])
@@ -135,7 +135,7 @@ class DummyTest < Test::Unit::TestCase
     end
   end
 
-  sub_test_case 'when dummy plugin has storage which is specified the path'  do
+  sub_test_case 'when sample plugin has storage which is specified the path'  do
     setup do
       FileUtils.rm_rf(TEST_PLUGIN_STORAGE_PATH)
       FileUtils.mkdir_p(File.join(TEST_PLUGIN_STORAGE_PATH, 'json'))
@@ -144,9 +144,9 @@ class DummyTest < Test::Unit::TestCase
 
     config2 = {
       '@id' => 'test-02',
-      'tag' => 'dummy',
+      'tag' => 'sample',
       'rate' => '0',
-      'dummy' => '[{"x": 1, "y": "1"}, {"x": 2, "y": "2"}, {"x": 3, "y": "3"}]',
+      'sample' => '[{"x": 1, "y": "1"}, {"x": 2, "y": "2"}, {"x": 3, "y": "3"}]',
       'auto_increment_key' => 'id',
     }
     conf2 = config_element('ROOT', '', config2, [
